+++ packages/components-angular/src/app/examples/input-search-example.component.ts
@@ -1,0 +1,28 @@
+import { CUSTOM_ELEMENTS_SCHEMA, ChangeDetectionStrategy, Component } from '@angular/core';
+import { PorscheDesignSystemModule } from '@porsche-design-system/components-angular';
+
+@Component({
+  selector: 'page-input-search-example',
+  template: `
+    <form (submit)="onSubmit($event)">
+      <p-input-search name="some-name" label="Some Label"></p-input-search>
+      <p-button type="submit">Submit</p-button>
+      <p-button type="reset">Reset</p-button>
+    </form>
+
+    <p-text>Last submitted data: {{ lastSubmittedData }}</p-text>
+  `,
+  changeDetection: ChangeDetectionStrategy.OnPush,
+  standalone: true,
+  imports: [PorscheDesignSystemModule],
+  schemas: [CUSTOM_ELEMENTS_SCHEMA],
+})
+export class InputSearchExampleComponent {
+  lastSubmittedData: string = '';
+
+  onSubmit(e: Event): void {
+    e.preventDefault();
+    const formData = new FormData(e.target as HTMLFormElement);
+    this.lastSubmittedData = formData.get('some-name') as string;
+  }
+}