+++ packages/components/src/components/input-password/input-password-styles.ts
@@ -17,6 +17,7 @@ import {
 
 export const getComponentCss = (
   disabled: boolean,
+  loading: boolean,
   hideLabel: BreakpointCustomizable<boolean>,
   state: FormState,
   toggle: boolean,
@@ -25,7 +26,7 @@ export const getComponentCss = (
   theme: Theme
 ): string => {
   return getCss({
-    ...getFunctionalComponentInputBaseStyles(disabled, hideLabel, state, compact, readOnly, theme, {
+    ...getFunctionalComponentInputBaseStyles(disabled, loading, hideLabel, state, compact, readOnly, theme, {
       '&[type="text"]': {
         textOverflow: 'ellipsis',
       },